回退角色保存菜单权限

+++ macula-cloud-system/src/main/java/dev/macula/cloud/system/service/impl/SysRoleServiceImpl.java
@@ -24,7 +24,6 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.google.common.base.Joiner;
 import dev.macula.boot.constants.GlobalConstants;
 import dev.macula.boot.result.Option;
 import dev.macula.boot.starter.security.utils.SecurityUtils;
@@ -192,18 +191,15 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
      * 修改角色的资源权限
      *
      * @param roleId
-     * @param mapMenuIds
+     * @param menuIds
      * @return
      */
     @Override
     @Transactional
     @CacheEvict(cacheNames = "system", key = "'routes'")
-    public boolean updateRoleMenus(Long roleId, Map<String,List<Long>> mapMenuIds) {
-        // 删除当前页角色菜单
-        sysRoleMenuService.remove(new LambdaQueryWrapper<SysRoleMenu>()
-                .eq(SysRoleMenu::getRoleId, roleId)
-                .in(Objects.nonNull(mapMenuIds.get("curPage")), SysRoleMenu::getMenuId, mapMenuIds.get("curPage")));
-        List<Long> menuIds = mapMenuIds.get("curSel");
+    public boolean updateRoleMenus(Long roleId, List<Long> menuIds) {
+        // 删除角色菜单
+        sysRoleMenuService.remove(new LambdaQueryWrapper<SysRoleMenu>().eq(SysRoleMenu::getRoleId, roleId));
         // 新增角色菜单关系
         if (CollectionUtil.isNotEmpty(menuIds)) {
             List<SysRoleMenu> roleMenus = menuIds.stream()
@@ -211,20 +207,13 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
                     .collect(Collectors.toList());
             sysRoleMenuService.saveBatch(roleMenus);
         }
+
         // 删除角色权限
-        if(Objects.nonNull(mapMenuIds.get("curPage"))) {
-            sysRolePermissionService.remove(new LambdaQueryWrapper<SysRolePermission>()
-                    .eq(SysRolePermission::getRoleId, roleId)
-                    .inSql(Objects.nonNull(mapMenuIds.get("curPage")), SysRolePermission::getPermissionId,
-                            "select id from sys_permission where menu_id in (" + Joiner.on(",").join(mapMenuIds.get("curPage")) + ")"));
-        }
+        sysRolePermissionService.remove(new LambdaQueryWrapper<SysRolePermission>().eq(SysRolePermission::getRoleId, roleId));
         // 新增角色权限关系（用勾选的菜单ID获取权限，用权限ID和角色ID组装sys_role_permission表)
-        if(CollectionUtil.isEmpty(menuIds)){
-            return true;
-        }
         List<SysPermission> permList = sysPermissionService.list(
                 new LambdaQueryWrapper<SysPermission>()
-                        .in(SysPermission::getMenuId, menuIds));
+                        .in(menuIds != null, SysPermission::getMenuId, menuIds));
         if (CollectionUtil.isNotEmpty(permList)) {
             List<SysRolePermission> rolePerms = permList.stream().map(perm -> new SysRolePermission(roleId, perm.getId())).collect(Collectors.toList());
             sysRolePermissionService.saveBatch(rolePerms);
